+++ .gitignore
@@ -1,0 +1,28 @@
+
+AVC_robot/i0.ppm
+AVC_robot/makefile
+AVC_robot/robot.exe
+AVC_robot/robot.o
+AVC_robot/sfml-audio-2.dll
+AVC_robot/sfml-audio-d-2.dll
+AVC_robot/sfml-graphics-2.dll
+AVC_robot/sfml-graphics-d-2.dll
+AVC_robot/sfml-network-2.dll
+AVC_robot/sfml-network-d-2.dll
+AVC_robot/sfml-system-2.dll
+AVC_robot/sfml-system-d-2.dll
+AVC_robot/sfml-window-2.dll
+AVC_robot/sfml-window-d-2.dll
+AVC_server/cam.png
+AVC_server/sfml-audio-2.dll
+AVC_server/sfml-audio-d-2.dll
+AVC_server/sfml-graphics-2.dll
+AVC_server/sfml-graphics-d-2.dll
+AVC_server/sfml-network-2.dll
+AVC_server/sfml-network-d-2.dll
+AVC_server/sfml-system-2.dll
+AVC_server/sfml-system-d-2.dll
+AVC_server/sfml-window-2.dll
+AVC_server/sfml-window-d-2.dll
+AVC_server/server3.o
+AVC_server/server3.exe
