--- AVC_robot/main.cpp
@@ -1,8 +1,0 @@
-#include "helper.hpp"
-#include "ImageProcessing.hpp"
-#include "Robot.hpp" 
-
-int main(){
-	Robot robot;
-	robot.start();
-}
